Code Refactor v0.1
