+++ ui/src/html_files/cibsterDash.html
