+++ ui/src/html_files/hosting.html
