+++ ui/src/html_files/questions.html
@@ -31,7 +31,7 @@ body, html {
 <!-- Navbar (sit on top) -->
 <div class="w3-top">
   <div class="w3-bar w3-black w3-card" id="myNavbar">
-    <a href="home.html" class="w3-bar-item w3-button w3-wide">LOGO</a>
+    <a href="/" class="w3-bar-item w3-button w3-wide">LOGO</a>
     <!-- Right-sided navbar links -->
     <div class="w3-right w3-hide-small">
       <a href="/compliments" class="w3-bar-item w3-button">COMPLIMENTS</a>
