+++ ui/src/html_files/restaurant.html
