+++ ui/src/html_files/send_email.html
