+++ ui/src/html_files/terminal.html
